add enemy abstraction
