+++ Assets/Scripts/Enemy/BaseEnemy.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cbd6092f5c3e18849a838777f1fc2ec5
+guid: dea671990d7726c4d905d5426721d12a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
